prayers now use the command queue as well, however we moved some client side code to the server side by mistake
